reorganize unicode into top level module
