+++ harfrust/Cargo.toml
@@ -17,12 +17,14 @@ bitflags = "2.9"
 bytemuck = { version = "1.22", features = ["extern_crate_alloc"] }
 core_maths = { version = "0.1", optional = true }
 smallvec = "1.14"
+icu_properties = { version = "2.2.0", optional = true }
 
 [features]
 default = ["std"]
 std = ["read-fonts/std"]
 libm = ["dep:core_maths", "read-fonts/libm"]
 experimental_font_api = []
+icu = ["dep:icu_properties"]
 
 [dev-dependencies]
 harfrust = { path = "../harfrust", features = ["experimental_font_api"] }
