+++ harfrust/src/algs.rs
