+++ harfrust/src/lib.rs
@@ -10,7 +10,9 @@ A complete [harfbuzz](https://github.com/harfbuzz/harfbuzz) shaping algorithm po
 
 extern crate alloc;
 
+mod algs;
 mod hb;
+mod unicode;
 
 #[cfg(feature = "std")]
 pub(crate) type U32Set = read_fonts::collections::int_set::U32Set;
